+++ packages/api-reference/src/features/Operation/Webhooks.vue
@@ -1,4 +1,5 @@
 <script lang="ts" setup>
+import { Lazy } from '@/components/Content/Lazy'
 import {
   CompactSection,
   Section,
@@ -12,8 +13,7 @@ import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { Webhooks } from '@scalar/types/legacy'
 import { computed } from 'vue'
 
-import { Lazy } from '../../../components/Content/Lazy'
-import Webhook from './Webhook.vue'
+import Webhook from './components/Webhook.vue'
 
 const props = defineProps<{
   webhooks: Webhooks
