+++ packages/api-reference/src/features/Operation/components/ContentTypeSelect.vue
