+++ packages/api-reference/src/features/Operation/components/OperationDescription.vue
