+++ packages/api-reference/src/features/Operation/components/OperationParameters.vue
