+++ packages/api-reference/src/features/Operation/components/OperationPath.vue
